requirements.txt
a2a==0.44
a2a-json-rpc==0.1.3
a2a-sdk==0.2.15
a2a-server==0.5.4
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
audioop-lts==0.2.2
Authlib==1.6.0
Automat==25.4.16
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chuk-ai-session-manager==0.7.1
chuk-llm==0.8.13
chuk-mcp==0.5
chuk-sessions==0.4.2
chuk-tool-processor==0.6.2
click==8.2.1
cloudpickle==3.1.1
constantly==23.10.4
contourpy==1.3.3
cryptography==45.0.4
cssselect==1.3.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
ecos==2.0.14
eval_type_backport==0.2.2
fastapi==0.115.14
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.5.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.100.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.22.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
groq==0.30.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
hyperlink==21.0.0
ibm-cos-sdk==2.14.2
ibm-cos-sdk-core==2.14.2
ibm-cos-sdk-s3transfer==2.14.2
ibm_watsonx_ai==1.3.31
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
iso8601==2.1.0
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
litellm==1.73.2
llvmlite==0.44.0
lomond==0.3.3
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mcp==1.10.1
mistralai==1.9.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numexpr==2.11.0
numpy==2.2.6
ollama==0.5.1
openai==1.93.0
openai-whisper==20250625
opentelemetry-api==1.35.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-exporter-prometheus==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
osqp==0.6.7.post3
packaging==25.0
pandas==2.2.3
parsel==1.10.0
pillow==11.3.0
prometheus_client==0.22.1
propcache==0.3.2
Protego==0.5.0
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdldl==0.1.7.post5
queuelib==1.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scikit-survival==0.24.1
scipy==1.16.0
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sse-starlette==2.3.6
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
streamlit==1.47.1
streamlit_mic_recorder==0.0.8
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.8.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.2
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uuid==1.30
uvicorn==0.34.3
w3lib==2.3.1
watchdog==6.0.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
